Introduce utils Module to Reuse Common Code
