+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Changes
 
-## Version 1.0.0 (Not yet Release)
+## Version 1.0.0 (Not yet Released)
 
+* Create a New Maven Module "utils" for Common Code Reuse - Issue #720
 * Implement ReplicationStateImpl to Manage and Cache Token Range to Replica Mappings - Issue #719
 * Implement NodeResolverImpl to Resolve Nodes by IP Address and UUID - Issue #718
 * Specify Interval for Next Connection - Issue #674
