+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/providers/AgentNativeConnectionProvider.java
@@ -86,7 +86,7 @@ public class AgentNativeConnectionProvider implements DistributedNativeConnectio
         DistributedNativeBuilder nativeConnectionBuilder =
                 DistributedNativeConnectionProviderImpl.builder()
                         .withInitialContactPoints(resolveInitialContactPoints(agentConnectionConfig.getContactPoints()))
-                        .withAgentType(agentConnectionConfig.getType().toString())
+                        .withAgentType(agentConnectionConfig.getType())
                         .withLocalDatacenter(agentConnectionConfig.getLocalDatacenter())
                         .withAuthProvider(authProvider)
                         .withSslEngineFactory(sslEngineFactory);
