+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/spring/BeanConfigurator.java
@@ -24,9 +24,10 @@ import com.ericsson.bss.cassandra.ecchronos.core.impl.metadata.NodeResolverImpl;
 import com.ericsson.bss.cassandra.ecchronos.core.impl.state.ReplicationStateImpl;
 import com.ericsson.bss.cassandra.ecchronos.core.metadata.NodeResolver;
 import com.ericsson.bss.cassandra.ecchronos.core.state.ReplicationState;
-import com.ericsson.bss.cassandra.ecchronos.data.exceptions.EcChronosException;
 import com.ericsson.bss.cassandra.ecchronos.data.sync.EccNodesSync;
 
+import com.ericsson.bss.cassandra.ecchronos.utils.exceptions.ConfigurationException;
+import com.ericsson.bss.cassandra.ecchronos.utils.exceptions.EcChronosException;
 import java.net.InetAddress;
 import java.io.IOException;
 import java.net.UnknownHostException;
@@ -38,7 +39,6 @@ import com.ericsson.bss.cassandra.ecchronos.application.config.Config;
 import com.ericsson.bss.cassandra.ecchronos.application.config.ConfigRefresher;
 import com.ericsson.bss.cassandra.ecchronos.application.config.ConfigurationHelper;
 import com.ericsson.bss.cassandra.ecchronos.application.config.security.Security;
-import com.ericsson.bss.cassandra.ecchronos.application.exceptions.ConfigurationException;
 import com.ericsson.bss.cassandra.ecchronos.application.providers.AgentNativeConnectionProvider;
 import com.ericsson.bss.cassandra.ecchronos.connection.CertificateHandler;
 import com.ericsson.bss.cassandra.ecchronos.connection.DistributedNativeConnectionProvider;
