+++ core/pom.xml
@@ -28,6 +28,13 @@
     <artifactId>core</artifactId>
 
     <dependencies>
+        <!-- Internal -->
+        <dependency>
+            <groupId>com.ericsson.bss.cassandra.ecchronos</groupId>
+            <artifactId>utils</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
         <!-- Cassandra driver -->
         <dependency>
             <groupId>com.datastax.oss</groupId>
