+++ utils/src/main/java/com/ericsson/bss/cassandra/ecchronos/utils/enums/package-info.java
@@ -12,28 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ericsson.bss.cassandra.ecchronos.application.exceptions;
-
-public class ConfigurationException extends Exception
-{
-    private static final long serialVersionUID = -2269440899665538081L;
-
-    public ConfigurationException()
-    {
-    }
-
-    public ConfigurationException(final String message)
-    {
-        super(message);
-    }
-
-    public ConfigurationException(final String message, final Throwable cause)
-    {
-        super(message, cause);
-    }
-
-    public ConfigurationException(final Throwable cause)
-    {
-        super(cause);
-    }
-}
+/**
+ * Contains the enums related with ecChronos operations.
+ */
+package com.ericsson.bss.cassandra.ecchronos.utils.enums;
