+++ utils/src/main/java/com/ericsson/bss/cassandra/ecchronos/utils/enums/sync/NodeStatus.java
@@ -12,7 +12,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.ericsson.bss.cassandra.ecchronos.utils.enums.sync;
+
 /**
- * Contains exceptions related to I/O in data tables.
+ * The status of nodes after creating jmx connection.
  */
-package com.ericsson.bss.cassandra.ecchronos.data.exceptions;
+public enum NodeStatus
+{
+    UNAVAILABLE,
+    AVAILABLE,
+    UNREACHABLE
+}
+
